requirements.txt
fastapi
uvicorn
timm
pillow
python-multipart